create thescore api module
